Fixed a bunch of issues
issues fixed

+++ Assets/ObstacleMovement.cs
@@ -7,11 +7,13 @@ public class ObstacleMovement : MonoBehaviour {
     const float speed = -10;
     public float laneWidth = 2;
     float X = 0;
-    int lane = 0;
+    float lane = 0;
 
 	// Use this for initialization
 	void Start () {
         lane = Random.Range(-1, 1);
+
+        
 
         float targetX = lane * laneWidth;
         X = (targetX - transform.position.x) * .1f;
